refactor: clean up node and comment imports
